This commit's primary goal is to restructure and separate the essential parts of the system, hence the branch name "rj/systemRefit". Throughout this commit, several server tasks were abstracted away from each other and given their own expressjs apps, "sub-apps" for the server to use. These sub tasks include (so far) running the test page, and handling all database transactions with a single app. Giving each task its own app allows for better structuring and modularity of the entire system. I also forsee that it will aid in system expansion throughout its lifetime. Below is a detailed list of changes:
public/js/test.js:
    - Modified all database interface endpoints to reflect the system refit (i.e. abstraction of all database functions to the "/mdbi" endpoint)

server.js:
    - Modified var logger and var handles to use the new /util directory path stored in settings.util
    - Added the server.js handlerTag for use with logging
    - Modified all logger.log() calls to use the server.js handlerTag
    - Removed all abstracted server routes (i.e. both the test page and database functions that are now abstracted away from the main server)
    - Added initialization and implementation of the test page sub-application, which is a sub-app that now handles all test page requests at the "/test" endpoint
    - Added initialization and implementation of the MDBI (MongoDB Interface) sub-application, which is a sub-app that now handles all database transactions

util/logger.js:
    - Added an initial element in logger.dataQueue to invoke automatic log file initialization writing
    - Added a space at the end of the sourceTag for better log aesthetic

util/mongoWrapper.js:
    - Moved the entire file to the new mdbi/ directory for better system modularity (i.e. part of database function abstraction)

util/route_handlers.js:
    - Modified var logger to use the new /util directory path stored in settings.util
    - Removed the ALL_COLLECTIONS constant for better system modularity (i.e. part of database function abstraction)
    - Removed var database for better system modularity (i.e. part of database function abstraction)
    - Removed MongoDB Client Instance for better system modularity (i.e. part of database function abstraction)
    - Removed all test page handler functions from handle_map for better system modularity (i.e. part of test page function abstraction)
    - Removed all database request handler functions from handle_map for better system modularity (i.e. part of database function abstraction)
    - Removed all utility methods used in conjunction with database functions (i.e. part of database function abstraction)

util/settings.js:
    - Added a file description for conformity's sake (since this file lacked any in-line documentation)

mdbi/*:
    - Moved mongoWrapper.js here
    - Added a README to describe the MDBI module
    - Added mdbi/app.js as the mdbi sub-app
    - Added mdbi/mdbiRoutes/index.js as the new container for all database functions (i.e. part of database function abstraction)

test/*:
    - Added test/app.js as the test page sub-app to handle all test page-related requests
    - Added test/testRoutes/index.js as the new container for all test page routes (i.e. part of test page function abstraction)
