+++ mdbi/README.txt
@@ -1,0 +1,5 @@
+=================
+MongoDB Interface
+=================
+
+    This directory houses the ExpressJS sub-app used by server.js (and all other sub-apps) to issue MongoDB database transactions. It was created within Core-v4's rj/systemRefit branch in an effort to abstract various functions away from each other, with the goal of better-structuring system software to a point where each system is individually maintainable and somewhat stand-alone.
